backend/requirements.txt
# Python requirements for FinSight backend
# Backend data fetching and NLP processing
# Generated with pip freeze

# Data fetching
requests==2.31.0
urllib3==2.1.0

# Data processing
pandas==2.1.4
numpy==1.26.2
lxml==5.0.0
openpyxl==3.1.2

# NLP processing
spacy==3.7.2
ja-core-news-md @ https://github.com/explosion/spacy-models/releases/download/ja_core_news_md-3.7.0/ja_core_news_md-3.7.0-py3-none-any.whl

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
